# ASCII optimization check in toCaser, for asLowerCase and asUpperCase