This optimizes std.uni.asLowerCase and std.uni.asUpperCase by adding an ascii check in the shared std.uni.toCaser template. This is to help address issue: https://issues.dlang.org/show_bug.cgi?id=11229. There have been similar optimizations in other case conversion code recently to address this issue. This change tries to be consistent with the others.

In my tests, this change sped up asLowerCase in latin languages about 75% (English, German, Spanish, Finnish). For example, 1000 repetitions over a Finnish text went from 23 seconds to 5 seconds. Similarly, a German text from 31 seconds to 7 seconds. Japanese and Chinese texts were unchanged, as expected. Tests were done on OSX with dmd settings `-release -O -boundscheck=off -inline`.

Note: This is my first pull request, please review appropriately.