@jondegenhardt 2016-05-10
Thanks for the quick review, fixed the else formatting.