@DmitryOlshansky 2016-05-10
Auto-merge toggled on